+++ apps/chaingraph-frontend/src/components/flow/nodes/ChaingraphNode/SchemaEditor/SchemaEditor.tsx
@@ -440,6 +440,7 @@ function AddPropertyDialog({
             title: 'Item',
           },
           isMutable: true,
+          isSchemaMutable: true,
         } as ArrayPortConfig
         break