+++ packages/chaingraph-types/src/port/instances/AnyPort.ts
@@ -8,8 +8,6 @@
 
 import type { JSONValue } from '../../utils/json'
 import type { AnyPortConfig, AnyPortValue, IPort, IPortConfig } from '../base'
-import { PortFactory } from '../../port/factory'
-import { deepCopy } from '../../utils'
 import { BasePort } from '../base'
 import { generatePortID } from '../id-generate'
 import { AnyPortPlugin } from '../plugins'
@@ -106,6 +104,7 @@ export class AnyPort extends BasePort<AnyPortConfig> {
       super.setValue(newValue)
       return
     }
+
     // Clear the current value if newValue is undefined
     if (newValue === undefined) {
       this.value = undefined
@@ -304,45 +303,6 @@ export class AnyPort extends BasePort<AnyPortConfig> {
       ...this.config,
       id: generatePortID(this.config.key || this.config.id || ''),
     })
-  }
-
-  operateOnUnderlyingType<UnderlyingTypeConfig extends IPortConfig = IPortConfig>(
-    operation: (underlyingPort: IPort<UnderlyingTypeConfig>) => IPort<UnderlyingTypeConfig>,
-  ): void {
-    const underlyingType = this.unwrapUnderlyingType() as UnderlyingTypeConfig
-    if (!underlyingType) {
-      throw new Error('No underlying type set for AnyPort')
-    }
-
-    const underlyingPort = PortFactory.createFromConfig({
-      ...deepCopy(underlyingType),
-      id: this.config.id,
-      parentId: this.config.parentId,
-      key: this.config.key,
-      title: this.config.title,
-      description: this.config.description,
-      connections: this.config.connections,
-      order: this.config.order,
-      nodeId: this.config.nodeId,
-      direction: this.config.direction,
-    }) as IPort
-    if (!underlyingPort) {
-      throw new Error('Failed to create underlying port from configuration')
-    }
-
-    underlyingPort.setValue(this.getValue())
-
-    // TODO: Find a way to type this properly
-    const changedPort = operation(underlyingPort as unknown as IPort<UnderlyingTypeConfig>)
-    if (!changedPort) {
-      throw new Error('Operation on underlying port returned undefined')
-    }
-
-    // Update the underlying type in the any port configuration
-    this.setUnderlyingType(changedPort.getConfig())
-
-    // Update the value if it exists
-    this.setValue(changedPort.getValue())
   }
 }